Added authentication to task endpoints

+++ src/routers/task.js
@@ -25,8 +25,8 @@ router.post('/tasks', auth, async (req, res) => {
     }
 })
 
-router.get('/tasks', async (req, res) => {
-    tasks = await Task.find({})
+router.get('/tasks', auth, async (req, res) => {
+    tasks = await Task.find({ owner: req.user._id })
     try {
         res.send(tasks)
     }
@@ -35,10 +35,12 @@ router.get('/tasks', async (req, res) => {
     }
 })
 
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id
+    const user = req.user
     try {
-        task = await Task.findById(_id)
+        // const task = await Task.findById(_id)
+        const task = await Task.findOne({ _id, owner: user._id })
         if(!task)
             return res.status(404).send()
         res.send(task)
@@ -48,19 +50,18 @@ router.get('/tasks/:id', async (req, res) => {
     }
 })
 
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['description', 'completed']
     const isAllowedUpdate = updates.every((update) => allowedUpdates.includes(update))
     if(!isAllowedUpdate) return res.status(400).send('Invalid update!')
     
     try {
-        const task = await Task.findById(req.params.id)
-
-        updates.forEach((update) => task[update] = req.body[update])
-        await task.save()
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id })
 
         if(!task) return res.status(404).send('Invalid task!')
+        updates.forEach((update) => task[update] = req.body[update])
+        await task.save()
 
         res.send(task)
     }
@@ -69,9 +70,9 @@ router.patch('/tasks/:id', async (req, res) => {
     }
 })
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findOneAndDelete(req.params.id)
+        const task = await Task.findOneAndDelete({ _id: req.params.id, owner: req.user._id })
         if(!task) return res.status(404).send('No task deleted')
 
         res.send(task)
